requirements.txt
celery
python-dotenv
redis
watchdog
fastapi
uvicorn